\name{LoadFn}
\alias{LoadFn}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{LoadFn}
\description{}
\usage{
LoadFn(directory, mods,...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{file}{file to load}
  \item{\dots}{Other arguments can be passed.}  
}
\value{}
\seealso{}
